Created list component
