+++ styles.css
@@ -46,6 +46,17 @@ html, head, body, #root {
     color: var(--dark-top)
 }
 
+.list {
+    display: flex;
+
+    width: 100%;
+    height: 100%;
+    
+    background-color: var(--dark-1);
+
+    color: var(--dark-top)
+}
+
 /* Sidebar */
 
 .sidebar {
